pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cr-agent"
version = "0.1.0"
description = "AI code reviewer — review PRs and local diffs from your terminal"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [{ name = "Tanay Jagadeesh" }]
keywords = ["code-review", "ai", "cli", "llm", "github"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Quality Assurance",
]

dependencies = [
    "click>=8.1",
    "rich>=13.0",
    "openai>=1.0",
    "anthropic>=0.30",
    "requests>=2.31",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "tomli>=2.0; python_version < '3.12'",
    "tomli-w>=1.0",
]

[project.optional-dependencies]
server = [
    "fastapi",
    "uvicorn[standard]",
    "sqlalchemy[asyncio]",
    "aiosqlite",
    "alembic",
    "pydantic-settings",
    "httpx",
    "pygithub",
]

[project.scripts]
cr = "backend.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["backend", "prompts"]

[tool.hatch.build.targets.wheel.shared-data]
"prompts" = "share/cr-agent/prompts"
